feat: add pop out mode (#210)
Ref. Dintero/epics#230
Rel. Dintero/checkout#1107

Adds new functionality in the SDK where the checkout has both an embedded part but the actual payment flow is completed in a pop out. In order for this to work the button to launch the pop out is rendered client-side in the same runtime as the SDK. This is needed so that we can reliably launch and interact with the pop out from the SDK run-time. When the pop out is opened we also add a backdrop covering the entire screen for the current website. This should hinder the end user from changing the content of the session in the client while the user is completing the payment.

When the Checkout is opened in pop out mode:

- Payment button design is controlled by the embedded checkout
- Button and backdrop text is controlled by the embedded checkout
- Apple pay should work with all compatible safari browsers as it is not from inside an iframe
- Walley Bank ID flow is no longer embedded (might solve some issues for Swedish end useres)
- The Klarna and Google Pay backdrop will cover our entire payment pop out instead of just the embedded component (while our backdrop in turn covers the client page running the Dintero SDK)