fix: pop out loaded check different origin (#258)
Ref. Dintero/epics#230
Rel. Dintero/checkout#1170

Fixes logic when checking if the checkout pop out has loaded. When the SDK runtime is on a different origin than the checkout we are noe able to add event listeners to the pop out. Instead we now listen for a message posted by the checkout pop out JS runtime when the checkout application is opened so the SDK can know that the pop out has been loaded successfully.