fix(build): use exec plugin for github asset (#332)
Fixes #331
Ref. #330
Rel. semantic-release/npm#535
Rel. https://github.com/Dintero/Dintero.Checkout.Web.SDK/actions/runs/8045967660/job/21972265552

The previous config change did not work as intended. The npm plugin for semantic release seems to be running twice and the .tgz is not present in the dist folder, and from the logs it does not even look like the package created contains the actual build files either. Instead of trying to make sense of the npm plugin we try using the exec plugin to run npm pack before the github plugin creates is release.