fix(pop-out): relay refresh, lock and ppt messages (#369)
Ref. https://github.com/Dintero/epics/issues/445

Ensure that messages from the SDK checkout instance will also be relayed to the pop-out window. This change will make it so that if any of the `lockSession()`, `refreshSession()` or `setActivePaymentProductType()` functions are called after the pop out has been opened, it will also update the checkout in the pop-out window.

This change is somewhat related to the upcoming change to the checkout where we will make the checkout frontend fully support the `lockSession` and `refreshSession` in non-express mode.