fix: increase pop-out timeout
Increases the timeout for checkout not loaded from 3 seconds to 10. This primarily exists to have a viable fallback for allowing payment if for some reason we are unable to load the checkout (but still create the pop out window), or if it is closed immediately after it has been opened before we are able to attach event handlers to the close event.